codacy fixes
