+++ src/trie/node.go
@@ -2,10 +2,6 @@ package trie
 
 import (
 	"fmt"
-	"io"
-	"strings"
-
-	"github.com/golang/protobuf/proto"
 
 	"github.com/Fantom-foundation/go-lachesis/src/trie/internal"
 )
@@ -88,117 +84,4 @@ func (n hashNode) fstring(ind string) string {
 }
 func (n valueNode) fstring(ind string) string {
 	return fmt.Sprintf("%x ", []byte(n))
-}
-
-func mustDecodeNode(hash, buf []byte, cachegen uint16) node {
-	n, err := decodeNode(hash, buf, cachegen)
-	if err != nil {
-		panic(fmt.Sprintf("node %x: %v", hash, err))
-	}
-	return n
-}
-
-// decodeNode parses the protobuf encoding of a trie node.
-func decodeNode(hash, buf []byte, cachegen uint16) (node, error) {
-	if len(buf) == 0 {
-		return nil, io.ErrUnexpectedEOF
-	}
-
-	var x internal.Node
-	err := proto.Unmarshal(buf, &x)
-	if err != nil {
-		return nil, fmt.Errorf("decode error: %v", err)
-	}
-
-	switch x.Kind.(type) {
-	case *internal.Node_Short:
-		n, err := decodeShort(hash, x.GetShort(), cachegen)
-		return n, wrapError(err, "short")
-	case *internal.Node_Full:
-		n, err := decodeFull(hash, x.GetFull(), cachegen)
-		return n, wrapError(err, "full")
-	default:
-		return nil, fmt.Errorf("unexpected node: %+v", x.Kind)
-	}
-}
-
-func decodeShort(hash []byte, x *internal.ShortNode, cachegen uint16) (node, error) {
-	flag := nodeFlag{hash: hash, gen: cachegen}
-	key := compactToHex(x.Key)
-	if hasTerm(key) {
-		// value node
-		if v, ok := x.Val.Kind.(*internal.Node_Value); ok {
-			return &shortNode{key, valueNode(v.Value), flag}, nil
-		} else {
-			return nil, fmt.Errorf("invalid value node: %+v", x.Val.Kind)
-		}
-	}
-	r, err := decodeRef(x.Val, cachegen)
-	if err != nil {
-		return nil, wrapError(err, "val")
-	}
-	return &shortNode{key, r, flag}, nil
-}
-
-func decodeFull(hash []byte, x *internal.FullNode, cachegen uint16) (*fullNode, error) {
-	n := &fullNode{flags: nodeFlag{hash: hash, gen: cachegen}}
-	for i := 0; i < 16; i++ {
-		cld, err := decodeRef(x.Children[i], cachegen)
-		if err != nil {
-			return n, wrapError(err, fmt.Sprintf("[%d]", i))
-		}
-		n.Children[i] = cld
-	}
-
-	// value node
-	if v, ok := x.Children[16].Kind.(*internal.Node_Value); ok {
-		if len(v.Value) > 0 {
-			n.Children[16] = valueNode(v.Value)
-		}
-	} else {
-		panic(fmt.Errorf("unexpected node: %+v", x.Children[16].Kind))
-	}
-
-	return n, nil
-}
-
-func decodeRef(x *internal.Node, cachegen uint16) (node, error) {
-	switch x.Kind.(type) {
-	case *internal.Node_Hash:
-		return hashNode(x.GetHash()), nil
-	case *internal.Node_Value:
-		if len(x.GetValue()) > 0 {
-			return valueNode(x.GetValue()), nil
-		} else {
-			return nil, nil
-		}
-	case *internal.Node_Short:
-		return decodeShort(nil, x.GetShort(), cachegen)
-	case *internal.Node_Full:
-		return decodeFull(nil, x.GetFull(), cachegen)
-	default:
-		return nil, fmt.Errorf("unexpected node: %+v", x.Kind)
-	}
-}
-
-// wraps a decoding error with information about the path to the
-// invalid child node (for debugging encoding issues).
-type decodeError struct {
-	what  error
-	stack []string
-}
-
-func wrapError(err error, ctx string) error {
-	if err == nil {
-		return nil
-	}
-	if decErr, ok := err.(*decodeError); ok {
-		decErr.stack = append(decErr.stack, ctx)
-		return decErr
-	}
-	return &decodeError{err, []string{ctx}}
-}
-
-func (err *decodeError) Error() string {
-	return fmt.Sprintf("%v (decode path: %s)", err.what, strings.Join(err.stack, "<-"))
 }
